requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
click==8.1.8
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
ruff==0.9.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
